% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TTC.R
\docType{data}
\name{data_TTC}
\alias{data_TTC}
\title{Test dataset for the TTC package}
\format{A data frame with 5 rows and 1 variable:
\describe{
  \item{text}{Some text from Lorem Ipsum}
}}
\source{
\url{https://www.lipsum.com/feed/html}
}
\usage{
data_TTC
}
\description{
A dataset containing 5 strings.
}
\keyword{datasets}
