% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_words.R
\name{filter_words}
\alias{filter_words}
\title{Filtering words}
\usage{
filter_words(data, match_data = NULL, low_tresh = 0.005)
}
\arguments{
\item{data}{data is here}

\item{match_data}{shall be a integer of class character or numeric}

\item{low_tresh}{lower threshold. Do not recommend to lower this much more due to computational reasons.}
}
\value{
Returns the filtered words.
}
\description{
This is a complementary function to fix_the_string. Filter the data with another string of words or just a threshold.
}
