% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTC.R, R/TTC_info.R
\docType{package}
\name{TTC}
\alias{TTC}
\alias{TTC}
\alias{TTC-package}
\title{Takes a string and converts it to columns.}
\usage{
TTC(data_string, minSup, check_words = NULL)
}
\arguments{
\item{data_string}{The vaible/ string that you want to columns. Must be of class "character"}

\item{minSup}{minimum support for the words.}

\item{check_words}{a vector of words if you want to filter away some words manually. A pre-build in vector is in place for a small amount of common english words. Default is NULL.}
}
\value{
Returns a data frame with class "TTC".
}
\description{
The input is a variable of strings. The function filter the data and gets the most frequent words in the variable into a vaiable. This variable is the filtered with a choosen string or just a threshold.

The function "fix_the_string" is the function that has all the other functions embedded.
}
\section{TTC functions}{

fix_the_string
freq_word_tbl
words_to_col
filter_words
}

\section{TTC data}{

data_TTC
}

\examples{
data("data_TTC")
TTC(data_TTC$text, minSup = 0.01)


}
\author{
Eric Herwin
}
