% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_to_col.R
\name{words_to_col}
\alias{words_to_col}
\title{Words to column}
\usage{
words_to_col(word_string, str_ing_v, ...)
}
\arguments{
\item{word_string}{shall be a integer of class character or numeric}

\item{str_ing_v}{lower threshold. Do not recommend to lower this much more due to computational reasons}

\item{...}{Ignore. A complementary to fix_the_string.}
}
\value{
Returns the filtered words.
}
\description{
This is a complementary function to fix_the_string. Filter the data with another database or just a threshold.
}
