% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_the_string.R
\name{fix_the_string}
\alias{fix_the_string}
\title{Takes a string and converts it to columns.}
\usage{
fix_the_string(data_string, match_data = NULL, low_tresh = 0.005,
  seperator = c("punct", "digit"), parallel = FALSE)
}
\arguments{
\item{data_string}{The vaible/ string that you want to columns. Must be of class "character"}

\item{match_data}{An optional string to filter the words in the variable.}

\item{low_tresh}{Ignore if match_data is given. Sets a lower threshold to filter out non-frequent words porportinal to unique words in data_string.}

\item{seperator}{Default values "punct" and "digit", but accept more. See https://stat.ethz.ch/R-manual/R-devel/library/base/html/regex.html for more.}

\item{parallel}{Make the computation parallelized.}
}
\value{
Returns a data frame with class "TTC".
}
\description{
The input is a variable of strings. The function filter the data and gets the most frequent words in the variable into a vaiable. This variable is the filtered with a choosen string or just a threshold.
}
\examples{
data("data_TTC")
fix_the_string(data_TTC$text)

fix_the_string(data_TTC$text, parallel = TRUE)
}
