% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_the_string.R
\name{fix_the_string}
\alias{fix_the_string}
\title{Filtering words}
\usage{
fix_the_string(data_string, data_orginal, leng = 1000, match_data = NULL,
  low_tresh = 0.005)
}
\arguments{
\item{data_string}{more data}

\item{data_orginal}{shall be a integer of class character or numeric}

\item{leng}{desired length of the data}

\item{match_data}{Possible data base}

\item{low_tresh}{lower theshold}
}
\value{
Returns the merged data frame with columns of text
}
\description{
This is a complementary function to fix_the_string. Filter the data with another database or just a threshold.
}
