% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_word_tbl.R
\name{freq_word_tbl}
\alias{freq_word_tbl}
\title{Frequent words}
\usage{
freq_word_tbl(data_string, leng = 1000)
}
\arguments{
\item{data_string}{data string}

\item{leng}{length of data}
}
\value{
Returns the filtered words.
}
\description{
This is a complementary function to fix_the_string. Filter the data with another database or just a threshold.
}
