% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTC.R
\docType{package}
\name{TTC}
\alias{TTC}
\alias{TTC-package}
\title{TTC: A package that takes a variable with strings and converts it to columns.}
\description{
The function "fix_the_string" is the function that has all the other functions embedded.
}
\section{TTC functions}{

fix_the_string
freq_word_tbl
words_to_col
filter_words
}

\section{TTC data}{

data_TTC
}

\author{
Eric Herwin
}
