% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_to_col.R
\name{words_to_col}
\alias{words_to_col}
\title{Words to column}
\usage{
words_to_col(frequent_1_items, split_n_clean)
}
\arguments{
\item{frequent_1_items}{shall be a vector with thw frequency of the words}

\item{split_n_clean}{The splitted and cleaned words from freq_word_tbl}
}
\value{
Returns the filtered words.
}
\description{
This is a complementary function to TTC function.
}
